Update the framework
I create the Makefile that simplify to order the commands.
I create the lib folder to store the library and database of ROM and RAM.
I create the rtl folder to store the Systemverilog or verilog code.
I create the syn folder to store the files generated by Design Compiler.
I create the dat folder to store the patterns.
I create the report folder to store all the report about synthesis.
I create the scripts folder to store the synthesis and pt script.
I create the software folder to store the python code which can visualize
the pattern.